// src/enums.ts
export const enum MovieType {
    BOTH = '0',
    MOVIE = '1',
    SERIES = '2'
}

export const enum PosterType {
    UNSET = 0,
    POSTER = 1,
    COVER = 2,
    TEXTLESS = 3,
    LOGO = 4,
    PROMO = 5,
    CUSTOM = 6,
    ADVANCE = 7,
    TEASER = 8,
    THEATRICAL = 9,
    RE_RELEASE = 10,
    COMBO = 11,
    VIDEO_RELEASE = 12,
    CONCEPT = 13,
    NEVER_PRINTED = 14
}

export const enum GenreType {
    DRAMA = 1,
    ROMANCE = 2,
    SCI_FI = 3,
    THRILLER = 4,
    HORROR = 5
}

export const enum PosterThumbSizeType {
    XL = 'xl',
    L = 'l',
    S = 's',
    T = 't',
    M = 'm'
}

export const enum PersonThumbSizeType {
    S = 's',
    T = 't',
    M = 'm'
}

export const enum SortType {
    ORIGINAL_TITLE = 'original_title',
    NUM_POSTERS = 'num_posters',
    YEAR = 'year',
    RUNTIME = 'runtime',
    COUNTRY = 'country'
}

export const enum DirectionType {
    ASC = 'asc',
    DESC = 'desc'
}

export const enum Forms {
    SETTINGS = 'Settings'
}